remove legacy google pay functionality, remove lots of unused types
