+++ android/src/main/java/com/reactnativestripesdk/GooglePayButtonManager.kt
@@ -15,11 +15,6 @@ class GooglePayButtonManager : SimpleViewManager<GooglePayButtonView?>() {
     view.initialize()
   }
 
-  @ReactProp(name = "buttonType")
-  fun buttonType(view: GooglePayButtonView, buttonType: String) {
-    view.setButtonType(buttonType)
-  }
-
   @ReactProp(name = "type")
   fun type(view: GooglePayButtonView, buttonType: Int) {
     view.setType(buttonType)
