+++ src/NativeStripeSdk.tsx
@@ -2,7 +2,6 @@ import { NativeModules } from 'react-native';
 import type {
   PaymentMethod,
   PaymentIntent,
-  ApplePay,
   PlatformPay,
   PaymentSheet,
   SetupIntent,
@@ -18,10 +17,6 @@ import type {
   PresentPaymentSheetResult,
   ConfirmPaymentSheetPaymentResult,
   CreateTokenResult,
-  GooglePayInitResult,
-  PayWithGooglePayResult,
-  CreateGooglePayPaymentMethodResult,
-  GooglePay,
   OpenApplePaySetupResult,
   Token,
   VerifyMicrodepositsParams,
@@ -70,14 +65,6 @@ type NativeStripeSdkType = {
   createTokenForCVCUpdate(cvc: string): Promise<CreateTokenForCVCUpdateResult>;
   handleURLCallback(url: string): Promise<boolean>;
   createToken(params: Token.CreateParams): Promise<CreateTokenResult>;
-  isGooglePaySupported(params: GooglePay.IsSupportedParams): Promise<boolean>;
-  initGooglePay(params: GooglePay.InitParams): Promise<GooglePayInitResult>;
-  presentGooglePay(
-    params: GooglePay.PresentParams
-  ): Promise<PayWithGooglePayResult>;
-  createGooglePayPaymentMethod(
-    params: GooglePay.CreatePaymentMethodParams
-  ): Promise<CreateGooglePayPaymentMethodResult>;
   openApplePaySetup(): Promise<OpenApplePaySetupResult>;
   verifyMicrodeposits(
     isPaymentIntent: boolean,
@@ -104,7 +91,7 @@ type NativeStripeSdkType = {
   ): Promise<FinancialConnections.SessionResult>;
   resetPaymentSheetCustomer(): Promise<null>;
   isPlatformPaySupported(params: {
-    googlePay?: GooglePay.IsSupportedParams;
+    googlePay?: PlatformPay.IsGooglePaySupportedParams;
   }): Promise<boolean>;
   createPlatformPayPaymentMethod(
     params: PlatformPay.PaymentMethodParams,
@@ -112,8 +99,8 @@ type NativeStripeSdkType = {
   ): Promise<PlatformPay.PaymentMethodResult | PlatformPay.TokenResult>;
   dismissPlatformPay(): Promise<boolean>;
   updatePlatformPaySheet(
-    summaryItems: Array<ApplePay.CartSummaryItem>,
-    shippingMethods: Array<ApplePay.ShippingMethod>,
+    summaryItems: Array<PlatformPay.CartSummaryItem>,
+    shippingMethods: Array<PlatformPay.ShippingMethod>,
     errors: Array<PlatformPay.ApplePaySheetError>
   ): Promise<void>;
   confirmPlatformPay(
