+++ src/components/GooglePayButtonNative.tsx
@@ -1,5 +1,5 @@
 import { requireNativeComponent } from 'react-native';
-import type { GooglePayButtonComponent } from '../types';
+import type * as GooglePayButtonComponent from '../types/components/GooglePayButtonComponent';
 const GooglePayButtonNative =
   requireNativeComponent<GooglePayButtonComponent.NativeProps>(
     'GooglePayButton'
