+++ src/types/ApplePay.ts
@@ -1,64 +1,3 @@
-export type ContactFieldsType =
-  | 'emailAddress'
-  | 'name'
-  | 'phoneNumber'
-  | 'phoneticName'
-  | 'postalAddress';
-
-export type AddressFields =
-  | 'street'
-  | 'city'
-  | 'subAdministrativeArea'
-  | 'state'
-  | 'postalCode'
-  | 'country'
-  | 'countryCode'
-  | 'subLocality';
-
-export interface ShippingMethod {
-  /** A short, localized description. */
-  label: string;
-  /** The cost associated with this shipping option. */
-  amount: string;
-  /** When creating items for estimates or charges whose final value is not yet known, set this to true. */
-  isPending?: boolean;
-  /** A unique identifier for the shipping method. */
-  identifier: string;
-  /** A user-readable description of the shipping method. For example “Ships in 24 hours.” Don't repeat the content of the 'label' property. */
-  detail?: string;
-  /** The unix timestamp of the start date of the expected range of delivery or shipping dates for a package, or the time range when an item is available for pickup. Measured in seconds. */
-  startDate?: number;
-  /** The unix timestamp of the end date of the expected range of delivery or shipping dates for a package, or the time range when an item is available for pickup. Measured in seconds. */
-  endDate?: number;
-}
-
-interface PostalAddress {
-  city?: string;
-  country?: string;
-  postalCode?: string;
-  state?: string;
-  street?: string;
-  isoCountryCode?: string;
-  subAdministrativeArea?: string;
-  subLocality?: string;
-}
-
-interface ContactName {
-  familyName?: string;
-  namePrefix?: string;
-  nameSuffix?: string;
-  givenName?: string;
-  middleName?: string;
-  nickname?: string;
-}
-
-export interface ShippingContact {
-  emailAddress?: string;
-  name: ContactName;
-  phoneNumber?: string;
-  postalAddress: PostalAddress;
-}
-
 export type CartSummaryItem =
   | DeferredCartSummaryItem
   | ImmediateCartSummaryItem
@@ -98,13 +37,3 @@ export type RecurringCartSummaryItem = {
   label: string;
   amount: string;
 };
-
-export interface PresentParams {
-  cartItems: CartSummaryItem[];
-  country: string;
-  currency: string;
-  requiredShippingAddressFields?: ContactFieldsType[];
-  requiredBillingContactFields?: ContactFieldsType[];
-  shippingMethods?: ShippingMethod[];
-  jcbEnabled?: boolean;
-}
