+++ src/types/components/GooglePayButtonComponent.ts
@@ -2,5 +2,4 @@ import type { StyleProp, ViewStyle } from 'react-native';
 export interface NativeProps {
   style?: StyleProp<ViewStyle>;
   type?: number;
-  buttonType?: string;
 }
